+++ express/code/blocks/frictionless-quick-action-mobile/frictionless-quick-action-mobile.js
@@ -5,6 +5,7 @@ import {
   getIconElementDeprecated,
 } from '../../scripts/utils.js';
 import { transformLinkToAnimation } from '../../scripts/utils/media.js';
+import { createLocaleDropdownWrapper } from '../../scripts/widgets/locale-dropdown.js';
 
 let replaceKey; let getConfig;
 let loadScript;
@@ -107,94 +108,14 @@ function selectElementByTagPrefix(p) {
 const downloadKey = 'download-to-phone';
 const editKey = 'edit-in-adobe-express-for-free';
 
-function createLocaleDropdown() {
-  const dropdown = createTag('div', { class: 'custom-dropdown' });
-
-  // Create the dropdown button
-  const dropdownBtn = createTag('button', {
-    class: 'dropdown-btn',
-    type: 'button',
+function createCaptionLocaleDropdown() {
+  const { wrapper } = createLocaleDropdownWrapper({
+    label: 'Language spoken in video',
+    defaultValue: 'en-us',
+    onChange: (code) => {
+      selectedVideoLanguage = code;
+    },
   });
-
-  const buttonText = createTag('span', { class: 'dropdown-text' }, 'English (US)');
-  const chevronIcon = createTag('span', { class: 'dropdown-chevron' }, '▼');
-
-  dropdownBtn.append(buttonText, chevronIcon);
-
-  // Create the dropdown content container
-  const dropdownContent = createTag('div', { class: 'dropdown-content' });
-
-  const locales = [
-    { code: 'zh-hk', label: 'Chinese (Hong Kong)' },
-    { code: 'cmn-hans', label: 'Chinese (Simplified)' },
-    { code: 'cmn-hant', label: 'Chinese (Traditional)' },
-    { code: 'da-dk', label: 'Danish' },
-    { code: 'nl-nl', label: 'Dutch' },
-    { code: 'en-gb', label: 'English (UK)' },
-    { code: 'en-us', label: 'English (US)' },
-    { code: 'fr-fr', label: 'French' },
-    { code: 'de-de', label: 'German' },
-    { code: 'hi-in', label: 'Hindi' },
-    { code: 'it-it', label: 'Italian' },
-    { code: 'ja-jp', label: 'Japanese' },
-    { code: 'ko-kr', label: 'Korean' },
-    { code: 'nb-no', label: 'Norwegian' },
-    { code: 'pt-pt', label: 'Portuguese' },
-    { code: 'ru-ru', label: 'Russian' },
-    { code: 'es-es', label: 'Spanish' },
-    { code: 'sv-se', label: 'Swedish' },
-  ];
-
-  // Add all locale options
-  locales.forEach((locale) => {
-    const option = createTag('a', {
-      href: '#',
-      class: 'dropdown-option',
-      'data-value': locale.code,
-    }, locale.label);
-
-    option.addEventListener('click', (e) => {
-      e.preventDefault();
-      buttonText.textContent = locale.label;
-      selectedVideoLanguage = locale.code;
-      dropdownContent.classList.remove('show');
-      chevronIcon.style.transform = 'rotate(0deg)';
-    });
-
-    dropdownContent.append(option);
-  });
-
-  // Toggle dropdown on button click
-  dropdownBtn.addEventListener('click', (e) => {
-    e.preventDefault();
-    dropdownContent.classList.toggle('show');
-    const isOpen = dropdownContent.classList.contains('show');
-    chevronIcon.style.transform = isOpen ? 'rotate(180deg)' : 'rotate(0deg)';
-  });
-
-  // Close dropdown when clicking outside
-  document.addEventListener('click', (e) => {
-    if (!dropdown.contains(e.target)) {
-      dropdownContent.classList.remove('show');
-      chevronIcon.style.transform = 'rotate(0deg)';
-    }
-  });
-
-  dropdown.append(dropdownBtn, dropdownContent);
-  return dropdown;
-}
-
-function createLocaleDropdownWrapper() {
-  const wrapper = createTag('div', { class: 'locale-dropdown-wrapper' });
-
-  // Create label
-  const label = createTag('label', {
-    for: 'locale-select',
-    class: 'locale-dropdown-label',
-  }, 'Language spoken in video');
-
-  const localeDropdown = createLocaleDropdown();
-  wrapper.append(label, localeDropdown);
   return wrapper;
 }
 
@@ -521,7 +442,7 @@ export default async function decorate(block) {
 
     // Add locale dropdown for caption-video
     if (quickAction === 'caption-video') {
-      localeDropdownWrapper = createLocaleDropdownWrapper();
+      localeDropdownWrapper = createCaptionLocaleDropdown();
       dropzoneContainer.before(localeDropdownWrapper);
     }
   };
