+++ express/code/scripts/widgets/locale-dropdown.css
@@ -1,0 +1,148 @@
+/* Locale Dropdown Widget Styles */
+.locale-dropdown-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 400px;
+  margin: 24px auto 16px auto;
+  padding: 0 16px;
+  box-sizing: border-box;
+}
+
+.locale-dropdown-label {
+  font-size: 1.1rem;
+  color: #2c2c2c;
+  margin-bottom: 12px;
+  font-weight: 600;
+  text-align: center;
+  letter-spacing: 0.02em;
+}
+
+.custom-dropdown {
+  position: relative;
+  display: inline-block;
+  width: 100%;
+  max-width: 340px;
+}
+
+.dropdown-btn {
+  background-color: #ffffff;
+  color: #2c2c2c;
+  padding: 14px 16px;
+  font-size: 1rem;
+  font-weight: 500;
+  border: 2px solid #e0e0e0;
+  border-radius: 12px;
+  cursor: pointer;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  transition: all 0.2s ease;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.06);
+  outline: none;
+  box-sizing: border-box;
+}
+
+.dropdown-btn:hover {
+  border-color: #5c5ce0;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  transform: translateY(-1px);
+}
+
+.dropdown-btn:focus {
+  border-color: #5c5ce0;
+  box-shadow: 0 0 0 4px rgba(92, 92, 224, 0.15), 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.dropdown-text {
+  flex: 1;
+  text-align: left;
+}
+
+.dropdown-chevron {
+  transition: transform 0.2s ease;
+  font-size: 0.8rem;
+  color: #666;
+}
+
+.dropdown-content {
+  display: none;
+  position: absolute;
+  top: 100%;
+  left: 0;
+  right: 0;
+  background-color: #ffffff;
+  border: 2px solid #e0e0e0;
+  border-top: none;
+  border-radius: 0 0 12px 12px;
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.12);
+  z-index: 1000;
+  max-height: 250px;
+  overflow-y: auto;
+  animation: fadeInDown 0.2s ease-out;
+}
+
+.dropdown-content.show {
+  display: block;
+}
+
+.dropdown-option {
+  color: #2c2c2c;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+  font-size: 1rem;
+  font-weight: 500;
+  transition: background-color 0.15s ease;
+  border: none;
+  cursor: pointer;
+}
+
+.dropdown-option:hover {
+  background-color: #f8f9ff;
+  color: #5c5ce0;
+}
+
+.dropdown-option:active {
+  background-color: #5c5ce0;
+  color: #ffffff;
+}
+
+@keyframes fadeInDown {
+  from {
+    opacity: 0;
+    transform: translateY(-8px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .locale-dropdown-wrapper {
+    margin: 20px auto 12px auto;
+    padding: 0 12px;
+  }
+
+  .custom-dropdown {
+    max-width: 100%;
+  }
+
+  .dropdown-btn {
+    padding: 12px 14px;
+    font-size: 0.95rem;
+  }
+
+  .locale-dropdown-label {
+    font-size: 1rem;
+    margin-bottom: 10px;
+  }
+
+  .dropdown-content {
+    max-height: 200px;
+  }
+}
